Allow compilation with clang: #1907
Was getting "use of overloaded operator '=' is ambiguous (with operand types 'ColXpr' (aka 'Block<Eigen::Matrix<double, -1, -1, 0, -1, -1>, internal::traits<Matrix<double, -1, -1, 0, -1, -1> >::RowsAtCompileTime, 1, !IsRowMajor>') and 'Eigen::Matrix<double, -1, 1, 0, -1, 1>')".

The ambiguity is between DenseBase<OtherDerived> and EigenBase<OtherDerived>.  This looks to be a bug in either clang or eigen.  The ambiguity is resolved by using the same MatrixXd::block() on both sides.  It looks ugly, but it gets the job done.

All tests pass.