Modify assertAlmostEqual in tests/subtractExposures.py
Unit tests that rely on assertAlmostEqual with the "places"
argument are notoriously sensitive to differences in platform
and compiler.  I am using them here to test whether two
numbers are the same to within some tolerance.  This ticket
modifies the default behavior to instead use the "delta"
arguement, equivalent to replacing

  self.assertAlmostEqual(skp1[nk][np], skp2[nk][np], 4)

with

  self.assertTrue(abs(skp1[nk][np]-skp2[nk][np]) < 10**-4)

with the actual change being

  self.assertAlmostEqual(skp1[nk][np], skp2[nk][np], delta=1e-4)