Deblender now uses Image instead of MaskedImage for templates
This is intended to save memory, and though it's not as effective as
originally hoped, yet provides a useful savings.  The variance plane of
the templates was completely unused, and the mask plane was only used for
debugging that no-one ever looked at (and has since been removed to save
the additional mask planes).