initial check in of new 'DipoleFitTask'
eventually to replace existing dipole measurement/classification
tasks.

This is an initial implementation of the new dipole measurement
algorithm as described in DMTN-007. It requires a new `lmfit`
python package (and by dependency, `scipy`).
The primary advancements are its use of the pre-subtraction
images to impose constraints on the dipole fit (including a possible
linear background gradient in those images).

Further updates TBD will include:

Things still TBD for this are:
1. check if there is a gradient in pre-sub images that needs to be fit,
	if not don't include those params
2. better estimate of starting parameters for gradient
3. better estimate for starting flux when there's a strong gradient
4. separate gradient parameters for pos- and neg- images?
5. thoroughly test on simulated images with gradients (as done previously without gradients)
6. make sure we're setting chi2 weights correctly from variance planes
7. need to return chisq/dof (DONE) and assess quality of fit, set "isDipole" flag.
8. need to correctly evaluate and fill signal-to-noise slot
9. need to correctly ignore flagged pixels -- can we use numpy's ma module?
10. only compute stats (including fit) for pixels within dipole footprint,
	not entire footprint bbox
11. add debugging (ds9) output
12. better logging and robust handling (raising) of exceptions