Use all keys in templateIdList[0] to help define templateId
Previously only the 'visit' key from templateIdList[0] was used
(and was also required)
in combination with the sensorRef.dataId to define the new template.
Now uses all keys in the templateIdList[0].dataId to specify
a revised dataId based on the sensorRef.dataId

This commit is intended to be backward compatible.