Add extra log.info about template, science PSF.
Explicitly state the effective FWHM of the template and science PSFs
this is a useful supplement to lines such as
"Matching Psf FWHM 5.76 -> 6.00 pix"
which implicitly almost contain the same information but only in
the direction of convolution.
These new lines separately provide the PSF FWHM for the template, science
irrespective of which is being convolved.  This provides a useful check
on the "Matching Psf ..." values.