Remove constraints in ModelPsfMatchTask
ModelPsfMatchTask previously required the user to supply the modelPsf
dimensions identical to the science psf dimensions and required the
science PSF be a KernelPsf. These requirements were removed to enable
matching of WarpedPsfs which vary in dimensions across an image.
The reference model PSF must still be a KernelPSF until resized
method exist for all PSFs.