Pad science PSF if too small for matching kernel in ModelPsfMatchTask
Add three config parameters to ModelPsfMatchTask that
enable automatic or manual padding of the science PSF before matching
to a model PSF. By default, doAutoPadPsf is set to True and will
automatically pad the PSF to the minimum size appropriate,
as defined by the config autoPadPsfTo (in units of the kernel size).
If doAutoPadPsf is set to False, science PSFs will be padded by the
padPsfBy number of pixels.