Make test_dipoleFitter.py more robust
This includes setting the random seed explicitly in the test dataset
realization code and allowing for more configs to be specified in the
call to the detectDipoleSources() function in utils.py.  In particular,
the positive and negative images used to make the difference image
must have different noise realizations, so must be seeded with different
values.

This also adds functionality in the detectDipoleSources() function in
utils.py to allow for the default binSize for background estimation to be
adjusted as the current default of 128 is not appropriate for test
images typically on the order of 100x100 pixels.  To avoid decreasing
the binSize by too much, a minimum binSize input parameter is provided.