Add bbox check before constructing dcrModel.
The coadd patch contains a large buffer of pixels on all sides, and the dcrModel requires that a small number of those get clipped to prevent artifacts at patch boundaries. This adds a check before trying to build the dcrModel to skip a patch if there will not be enough pixels remaining to build a model. In that case, the patch should be skipped in any case, since any valid data would be contained in neighboring patches anyway.