Turn off compatibility mode as the default.
Also set the default convolution mode to `convolveTemplate`.
This is the next step in our transition to the refactored image differencing,
which preserves the default functionality of always convolving the template.
With `forceCompatibility=False` we can use the new code to fix a few bugs that were
present in the old ImageDifferenceTask.

Update subtractImages config validation.