Force a larger epsfcn value for scipy
lmfit>1.0.2 coerces the data and independent variables to float64, but
our internal model images are float32, so the default step sizes are
too small and the fits fail. Changing our model images to double would
be slower and take up more memory; epsfcn>1e-13 seems to be enough to
get good fits: step length is proportional to sqrt(epsfcn).

If we switch to least_squares, we would instead set `diff_step` here.