Fix existing unit tests after clearing the template mask.
The modifications in `test_order_equal_images` are needed because there is no border padding of the template.
The changes in `test_symmetry` are needed to make the images actually symmetric. This should have been the case before, but after clearing the EDGE flag the `template_better.difference` image was entirely NANs due to the statistics calculation in decorrelation still needing a mask plane. This was fixed by copying the difference mask plane to the template before decorrelation.