Manually filter nans prior to fitting.
The lmfit nan_policy needs all the variables to be aligned, but here we are
fitting with the "x" grid as the independent variable, but "z" has a different
shape.  But the nan_policy="omit" option in lmfit simply does a filtering prior
to doing any fit, so here we zero out any nans and ensure the weights are 0 for
these values.