Rename (with deprecation) symbols for subtractImages calibration.
The new names reflect the possibility of getting PhotoCalibs and WCSs
from the same catalogs, though we're not doing that yet.

In addition, whether external calibrations are applied when 'run'
methods are called directly is now contingent on whether external
calibrations were passed in, not a config option; the config option
just controls whether those connections are active (and hence how
'run' is called) when being run as a PipelineTask.