Reraise partial-template subtraction failures as NoWorkFound
Add new config parameter minTemplateFractionForExpectedSuccess
with default value of 0.2: minimum for ci_hsc
and ci_imsim to succeed without special overrides.
If the template coverage is above the requiredTemplateFraction,
but below minTemplateFractionForExpectedSuccess, image subtraction
will be attempted, but any exceptions will be reraised as NoWorkFound.

Recomputing the template coverage is an OK expendature of time
because the Task has failed already.