Add diffim limiting magnitude to Task metadata.
This uses photoCalib and psf information from the science and template
exposures to estimate the limiting magnitude of the difference image.
The result is added to the Task metadata, which is then propagated into
diffim metrics alongside existing values.

The diffim limiting magnitude is useful for, e.g., selecting only
diaSources in regions with good template depth, because in general,
it will be equal to the science limiting magnitude if the
template is high quality, but it will be degraded if the
tempalte is poor.