Update AL deconvolution to share code with convolution case.
As noted on the ticket (https://rubinobs.atlassian.net/browse/DM-47642), the original motivation for the choice of basis functions for the deconvolution kernel appears to have been a misinterpretation of the cited paper. The simpler approach gives comparable results, is easier to modify, and does not give the illusion of a mathematical foundation.