Use FootprintMergeList instead of FootprintSet.merge
This avoids the (unresolved) bug in merge(), and leaves the positive
footprints unmerged in the output catalog.
It does require an API change to processResults, since we now have to use
SourceCatalogs instead of FootprintSets, but I think that's an overall
improvement (though it does make the `if doDeblend` branches rather messier).